// src/users/user-controller.interface.ts
import { NextFunction, Request, Response } from 'express';

export interface UsersControllerInterface {
    register: (req: Request, res: Response, next: NextFunction) => void;
    login: (req: Request, res: Response, next: NextFunction) => void;
    info: (req: Request, res: Response, next: NextFunction) => void;
}